version al 10112020

+++ src/main/java/movil/pos/producto/repository/entity/Impuesto.java
@@ -1,7 +1,5 @@
 package movil.pos.producto.repository.entity;
 
-import java.math.BigDecimal;
-
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -23,7 +21,7 @@ public class Impuesto {
 
     private String nombre;
 
-    private BigDecimal tasa;
+    private int tasa;
 
     private boolean activo;
